Update TaskSupportPass to perform all instrumentation and infer the clock signal.
Clock signal changes:

The clock signal is used to drive the signaltap instance and count
cycles.
It used to be specified by user, but different applications may use
differnt signals.
This commit infer the right clock signal from always blocks which
contains display tasks.
Among the senslists of those always blocks, we select the
most-frequently-used signal as the clock signal.

Instrumentation changes:

Now TaskSupportPass is responsible to instrument both the cycle counter
and the signaltap instance.