Add Logic2RegPass, which converts logic type to wire and reg
This pass is a small effort towards converting systemverilog to verilog,
because we have encountered several issues of their incompatibility in
other tools.

Input/output logic ports are converted to wires.
Logic declarations inside a module are converted to regs.

Logic2RegPass is not enabled by default since it is still not enough to
convert systemverilog to verilog (e.g. always_comb + if condition)